+++ src/main/java/ladder/model/Users.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.model;
 
 import java.util.List;
 import java.util.stream.Collectors;
