+++ src/test/java/ladder/LineTest.java
@@ -1,5 +1,7 @@
 package ladder;
 
+import ladder.model.LadderGenerateStrategy;
+import ladder.model.Line;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
